Refactor RockQueryParams Interface with Enhanced Documentation
- Add detailed JSDoc comments to RockQueryParams interface
- Improve type clarity and inline documentation for Rock API query parameters
- Move BaseIconButtonProps to a separate types file for IconButton component
