+++ app/primitives/button/icon-button.primitive.tsx
@@ -2,41 +2,8 @@ import React from "react";
 import { Link, type LinkProps } from "react-router-dom";
 import { Button } from "./button.primitive";
 import Icon from "../icon";
-import { icons } from "~/lib/icons";
 import { twMerge } from "tailwind-merge";
-
-type IconName = keyof typeof icons;
-
-type BaseIconButtonProps = {
-  /**
-   * The text content of the button
-   */
-  children: React.ReactNode;
-  /**
-   * Whether to show the rotating arrow icon
-   */
-  withRotatingArrow?: boolean;
-  /**
-   * Additional CSS classes to apply to the button
-   */
-  className?: string;
-  /**
-   * The icon name to use (defaults to arrowBack for rotating arrow)
-   */
-  iconName?: IconName;
-  /**
-   * The size of the icon (defaults to 20)
-   */
-  iconSize?: number;
-  /**
-   * Text and border color
-   */
-  color?: string;
-  /**
-   * Optional click handler
-   */
-  onClick?: () => void;
-};
+import { BaseIconButtonProps } from "./types";
 
 export type IconButtonProps = BaseIconButtonProps &
   (
